chore: remove `creator` module
