+++ kernel/src/process/mod.rs
@@ -1,13 +1,19 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-mod creator;
 mod manager;
 mod stack_frame;
 
-use crate::{mem::allocator::page_box::PageBox, tests, tss::TSS};
+use crate::{
+    mem::{allocator::page_box::PageBox, paging::pml4::PML4},
+    tests,
+    tss::TSS,
+};
 use common::constant::INTERRUPT_STACK;
 use stack_frame::StackFrame;
-use x86_64::{structures::paging::PageTable, PhysAddr, VirtAddr};
+use x86_64::{
+    structures::paging::{PageTable, PageTableFlags},
+    PhysAddr, VirtAddr,
+};
 
 pub fn init() {
     register_initial_interrupt_stack_table_addr();
@@ -42,7 +48,7 @@ pub struct Process {
 }
 impl Process {
     pub fn new(f: fn() -> !) -> Self {
-        let pml4 = creator::Pml4Creator::new().create();
+        let pml4 = Pml4Creator::new().create();
         let pml4_addr = pml4.phys_addr();
 
         Process {
@@ -68,5 +74,33 @@ impl Process {
         self.stack_frame
             .as_ref()
             .expect("Stack frame is not created")
+    }
+}
+
+pub struct Pml4Creator {
+    pml4: PageBox<PageTable>,
+}
+impl Pml4Creator {
+    pub fn new() -> Self {
+        Self {
+            pml4: PageBox::user(PageTable::new()),
+        }
+    }
+
+    pub fn create(mut self) -> PageBox<PageTable> {
+        self.enable_recursive_paging();
+        self.map_kernel_area();
+        self.pml4
+    }
+
+    fn enable_recursive_paging(&mut self) {
+        let a = self.pml4.phys_addr();
+        let f =
+            PageTableFlags::PRESENT | PageTableFlags::WRITABLE | PageTableFlags::USER_ACCESSIBLE;
+        self.pml4[511].set_addr(a, f);
+    }
+
+    fn map_kernel_area(&mut self) {
+        self.pml4[510] = PML4.lock().level_4_table()[510].clone();
     }
 }
